+++ Cargo.toml
@@ -2,7 +2,7 @@
 resolver = "2"
 members = [
     "dartfrog",
-    "chat",
+    "chat", "common",
 ]
 
 [profile.release]
