+++ dartfrog/Cargo.toml
@@ -10,6 +10,7 @@ kinode_process_lib = { git = "https://github.com/kinode-dao/process_lib", rev = 
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 wit-bindgen = "0.24.0"
+common = { path = "../common" }
 
 [features]
 prod = []
